refactor: remove duplications
