+++ src/fs/read.js
@@ -1,8 +1,9 @@
 import path from 'path';
 import fsPromises from 'fs/promises';
-import { __dirname, FILES_DIR_NAME, ERRORS } from './constants.js';
 import FsOperationFailedError from './error.js';
+import { getConstants } from '../constants.js';
 
+const { __dirname, FILES_DIR_NAME, ERRORS } = getConstants(import.meta.url);
 const FILE_NAME = 'fileToRead.txt'
 
 export const read = async () => {
