+++ src/modules/cjsToEsm.mjs
@@ -2,11 +2,10 @@ import path from 'path';
 import { readFile } from 'fs/promises';
 import { release, version } from 'os';
 import { createServer as createServerHttp } from 'http';
-import { fileURLToPath } from 'url';
 import './files/c.js';
+import { getConstants } from '../constants.js';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+const { __dirname, __filename } = getConstants(import.meta.url);
 
 const random = Math.random();
