+++ src/streams/write.js
@@ -1,13 +1,12 @@
 import path from 'path';
 import fs from 'fs';
-import { fileURLToPath } from 'url';
+import { getConstants } from '../constants.js';
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const DIR_NAME = 'files';
+const { __dirname, FILES_DIR_NAME } = getConstants(import.meta.url);
 const FILE_NAME = 'fileToWrite.txt';
 
 export const write = async () => {
-  const pathToFile = path.join(__dirname, DIR_NAME, FILE_NAME);
+  const pathToFile = path.join(__dirname, FILES_DIR_NAME, FILE_NAME);
   const writeStream = fs.createWriteStream(pathToFile);
 
   process.stdin.pipe(writeStream)
