+++ src/zip/zip.js
@@ -2,11 +2,9 @@ import path from 'path';
 import fs from 'fs';
 import { pipeline } from 'stream/promises';
 import { createGzip, createGunzip } from 'zlib';
-import { fileURLToPath } from 'url';
+import { getConstants } from '../constants.js';
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-const DIR_NAME = 'files';
+const { __dirname, FILES_DIR_NAME } = getConstants(import.meta.url);
 const FILE_NAME = 'fileToCompress.txt';
 const COMPRESSED_FILE_NAME = 'archive.gz'
 
@@ -19,8 +17,8 @@ export const options = {
 export const zip = async (option = options.compress) => {
   try {
     const { src, dest, createTransform } = option;
-    const readStream = fs.createReadStream(path.join(__dirname, DIR_NAME, src));
-    const writeStream = fs.createWriteStream(path.join(__dirname, DIR_NAME, dest));
+    const readStream = fs.createReadStream(path.join(__dirname, FILES_DIR_NAME, src));
+    const writeStream = fs.createWriteStream(path.join(__dirname, FILES_DIR_NAME, dest));
   
     await pipeline(readStream, createTransform.call(null), writeStream);
   } catch (err) {
